+++ src/main/java/io/github/poshjosh/ratelimiter/web/spring/WebRateLimiterContextSpring.java
@@ -3,9 +3,9 @@ package io.github.poshjosh.ratelimiter.web.spring;
 import io.github.poshjosh.ratelimiter.web.core.WebRateLimiterContext;
 import io.github.poshjosh.ratelimiter.web.spring.uri.ResourceInfoProviderSpring;
 
-public final class WebRateLimiterContextSpring {
+final class WebRateLimiterContextSpring {
     private WebRateLimiterContextSpring() {}
-    public static WebRateLimiterContext.Builder builder() {
+    static WebRateLimiterContext.Builder builder() {
         return WebRateLimiterContext.builder()
             .resourceInfoProvider(new ResourceInfoProviderSpring())
             .classesInPackageFinder(new ClassesInPackageFinderSpring());