+++ src/test/java/io/github/poshjosh/ratelimiter/web/spring/uri/ClassLevelResourcePathTest.java
@@ -1,0 +1,21 @@
+package io.github.poshjosh.ratelimiter.web.spring.uri;
+
+import io.github.poshjosh.ratelimiter.web.core.util.ResourcePath;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+class ClassLevelResourcePathTest extends AbstractResourcePathTestBase {
+
+    @Test
+    void shouldMatchStartOf() {
+        ResourcePath<String> resourcePath = givenResourcePath("/numbers");
+        assertThat( resourcePath.matches("/numbers")).isTrue();
+        assertThat( resourcePath.matches("/numbers/1")).isTrue();
+        assertThat( resourcePath.matches("/letters/a")).isFalse();
+    }
+
+    ResourcePath<String> givenResourcePath(String... uris) {
+        return new ClassLevelResourcePath(uris);
+    }
+}